requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
pytest
httpx
requests
aiohttp
bcrypt==4.0.1
python-jose[cryptography]